+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - `Ferrum::Browser#debug` opens headless session in the browser devtools frontend [#519]
 - `Ferrum::Frame#parent` returns the parent frame if this frame is nested [#523]
 - `Ferrum::Frame#frame_element` returns the element in which the window is embedded [#524]
+- `Ferrum::Page#start_screencast` starts sending frames to record screencast [#494]
+- `Ferrum::Page#stop_screencast` stops sending frames [#494]
 
 ### Changed